new test file for validators
